+++ app/Helpers/DateHelper.php
@@ -2,8 +2,9 @@
 
 namespace App\Helpers;
 
-use Carbon\Carbon;
+use function Safe\date;
 use function Safe\strtotime;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Auth;
 
@@ -176,7 +177,7 @@ class DateHelper
      * Return the day of the date according to the timezone of the user
      * like "Mon", or "Wed".
      *
-     * @param Carbon $date
+     * @param \Carbon\Carbon $date
      * @return string
      */
     public static function getShortDay($date): string
@@ -188,7 +189,7 @@ class DateHelper
      * Return a date according to the timezone of the user, in a short format
      * like "Oct 29".
      *
-     * @param Carbon $date
+     * @param \Carbon\Carbon $date
      * @return string
      */
     public static function getShortDateWithoutYear($date): string
@@ -200,7 +201,7 @@ class DateHelper
      * Return a date and the time according to the timezone of the user, in a short format
      * like "Oct 29, 1981 19:32".
      *
-     * @param Carbon $date
+     * @param \Carbon\Carbon $date
      * @return string
      */
     public static function getShortDateWithTime($date): string
@@ -224,12 +225,12 @@ class DateHelper
 
     /**
      * Add a given number of week/month/year to a date.
-     * @param Carbon $date      the start date
+     * @param \Carbon\Carbon $date      the start date
      * @param string $frequency week/month/year
      * @param int $number    the number of week/month/year to increment to
-     * @return Carbon
+     * @return \Carbon\Carbon
      */
-    public static function addTimeAccordingToFrequencyType(Carbon $date, string $frequency, int $number): Carbon
+    public static function addTimeAccordingToFrequencyType(\Carbon\Carbon $date, string $frequency, int $number): \Carbon\Carbon
     {
         switch ($frequency) {
             case 'week':