+++ app/Services/User/UpdateViewPreference.php
@@ -33,6 +33,7 @@ class UpdateViewPreference extends BaseService
         $this->validate($data);
 
         try {
+            /** @var User */
             $user = User::where('account_id', $data['account_id'])
                 ->findOrFail($data['user_id']);
         } catch (ModelNotFoundException $e) {